Move all Torch modules to a subpackage
We were mixing Torch modules and Thinc models. Move all Torch modules
to a subpackage.
