+++ curated_transformers/models/__init__.py
@@ -1,9 +1,5 @@
-from .activations import GeluNew
-from .attention import AttentionMask
-from .albert.encoder import AlbertEncoder
-from .bert.encoder import BertEncoder
 from .hf_loader import build_hf_encoder_loader_v1
-from .roberta.encoder import RobertaEncoder
+from .scalar_weight import build_scalar_weight_v1
 from .transformer_model import (
     build_bert_transformer_model_v1,
     build_camembert_transformer_model_v1,
@@ -12,4 +8,3 @@ from .transformer_model import (
     build_xlmr_transformer_model_v1,
 )
 from .with_strided_spans import build_with_strided_spans_v1
-from .scalar_weight import build_scalar_weight_v1
